--- a/docker-compose.yml
+++ b/docker-compose.yml
@@ -31,4 +31,3 @@ services:
 
   redis:
     image: redis:7-alpine
-