--- a/voxpop/views.py
+++ b/voxpop/views.py
@@ -110,7 +110,7 @@ def admin_voxpop(request, voxpop_id: UUID = None):
                         state=Question.State.DISCARDED,
                     ),
                 },
-                "idp": voxpop.organisation.idp
+                "idp": voxpop.organisation.idp,
             }
             return render(request, "voxpop/admin/voxpop.html", context)
     return render(request, "voxpop/admin/auth_error.html")
@@ -180,7 +180,7 @@ def index(request):
         if voxpop:
             context = {
                 "voxpop": voxpop,
-                "idp": voxpop.organisation.idp
+                "idp": voxpop.organisation.idp,
             }
     return render(request, "voxpop/index.html", context)