… requirements.txt
… voxpop/static/thumbs-up.svg
… voxpop/static/voxpop.css