This fixes the "too many connections" issue (and upgrades a bunch of packages, and improves on the almost non-existing docker compose setup).

Before this update running siege (https://github.com/JoeDog/siege) like so `siege --concurrent 230 --reps 5 http://localhost:8000/` results in a bunch of "too many connections" errors and the following siege result:

```
Transactions:                2625    hits
Availability:                  90.52 %
Elapsed time:                  19.58 secs
Data transferred:              58.29 MB
Response time:               1618.01 ms
Transaction rate:             134.07 trans/sec
Throughput:                     2.98 MB/sec
Concurrency:                  216.92
Successful transactions:     2625
Failed transactions:          275
Longest transaction:         4890.00 ms
Shortest transaction:          10.00 ms
```

After enabling the new connection pooling in django 5.1 we get no errors and this result:

```
Transactions:                3450    hits
Availability:                 100.00 %
Elapsed time:                   9.36 secs
Data transferred:              18.43 MB
Response time:                549.29 ms
Transaction rate:             368.59 trans/sec
Throughput:                     1.97 MB/sec
Concurrency:                  202.46
Successful transactions:     3450
Failed transactions:            0
Longest transaction:         4070.00 ms
Shortest transaction:           0.00 ms
```